monitor_mqtt/requirements.txt
paho-mqtt>=1.1,<1.3
pandas-schema
